+++ src/actions/public/readContract.ts
@@ -50,7 +50,7 @@ export type ReadContractParameters<
     | 'stateOverride'
   >
 > &
-  ContractFunctionParameters<abi, 'pure' | 'view', functionName, args>
+  ContractFunctionParameters<abi, 'pure' | 'view', functionName, args, boolean>
 
 export type ReadContractReturnType<
   abi extends Abi | readonly unknown[] = Abi,
@@ -124,7 +124,7 @@ export async function readContract<
     )({
       ...(rest as CallParameters),
       data: calldata,
-      to: address,
+      to: address!,
     })
     return decodeFunctionResult({
       abi,