+++ src/utils/index.test.ts
@@ -147,6 +147,8 @@ test('exports utils', () => {
       "parseUnits",
       "parseEther",
       "parseGwei",
+      "createNonceManager",
+      "nonceManager",
     ]
   `)
 })